requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
Automat==25.4.16
babel==2.16.0
beautifulsoup4==4.12.2
bleach==6.2.0
blis==1.3.0
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.3
cloudpathlib==0.22.0
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.1
cssselect==1.3.0
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.11
darkdetect==0.8.0
dateutils==0.6.12
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dm-tree==0.1.8
docstring_parser==0.16
elastic-transport==9.1.0
elasticsearch==9.1.1
et-xmlfile==1.1.0
etils==1.11.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.19.1
flatbuffers==25.1.24
fonttools==4.60.0
fqdn==1.5.1
frozendict==2.3.8
fsspec==2024.12.0
gast==0.6.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.35.1
hyperlink==21.0.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
incremental==24.7.2
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itemadapter==0.12.2
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.8.0
keras-nightly==3.8.0.dev2025012603
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==4.9.2
marisa-trie==1.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.1.17
mistune==3.1.0
ml-dtypes==0.4.1
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.13
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.8.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.3.3
oauthlib==3.2.2
openai==1.109.1
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
preshed==3.0.10
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
Protego==0.5.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.11.9
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
queuelib==1.8.0
referencing==0.36.2
regex==2023.6.3
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
Scrapy==2.13.3
Send2Trash==1.8.3
sequential==1.0.0
service-identity==24.2.0
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
sklearn==0.0.post5
smart_open==7.3.1
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
tb-nightly==2.19.0a20250126
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-estimator==2.13.0rc0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.13.0rc1
tensorflow-metadata==1.16.1
termcolor==2.5.0
terminado==0.18.1
textblob==0.17.1
tf_nightly==2.19.0.dev20250125
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tkinterdnd2==0.4.3
tldextract==5.3.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.4.2
tqdm==4.65.0
traitlets==5.14.3
transformers==4.56.2
tweety-ns @ https://github.com/mahrtayyab/tweety/archive/main.zip#sha256=3a15e03b358be6b9ff2d8f99e0ba58e3a87cf287c7f75498bff825d85b6a8c87
Twisted==25.5.0
typer==0.19.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.20
zipp==3.21.0
zope.interface==8.0
